# baremaps-server-studio/src/main/resources/studio-openapi.yaml
openapi: 3.0.3
info:
  title: Baremaps API
  description: Draft for a Baremaps API.
  version: ${project.version}
servers:
  - url: 'http://localhost:8080'
...
paths:
...
  '/styles/{styleId}':
    get:
      tags:
        - Styles
      summary: |-
        fetch a style by id
      description: |-
        Fetches the style with identifier `styleId`. The set of
        available styles can be retrieved at `/styles`.

        Not all styles are available in all style encodings.
      operationId: getStyle
      parameters:
        - $ref: '#/components/parameters/styleId'
      responses:
        '200':
          description: |-
            The style
          content:
            application/vnd.mapbox.style+json:
              schema:
                $ref: '#/components/schemas/mb-style'
        '404':
          description: |-
            style not found
    put:
      tags:
        - Styles
      summary: |-
        replace a style or add a new style
      description: |-
        Replace an existing style with the id `styleId`. If no
        such style exists, a new style with that id is added.
      operationId: updateStyle
      parameters:
        - $ref: '#/components/parameters/styleId'
      requestBody:
        description: |-
          A single style in one of the supported style encodings.
        content:
          application/vnd.mapbox.style+json:
            schema:
              $ref: '#/components/schemas/mb-style'
      responses:
        '204':
          description: |-
            style updated or created successfully
    delete:
      tags:
        - Styles
      summary: |-
        delete a style
      description: |-
        Delete an existing style with the id `styleId`. If no
        such style exists, an error is returned.
      operationId: deleteStyle
      parameters:
        - $ref: '#/components/parameters/styleId'
      responses:
        '204':
          description: |-
            style deleted
        '404':
          description: |-
            style not found
...
  '/tilesets/{tilesetId}':
    get:
      tags:
        - Tilesets
      summary: |-
        fetch a tileset by id
      description: |-
        Fetches the tileset with identifier `tilesetId`. The set of
        available tilesets can be retrieved at `/tilesets`.
      operationId: getTileset
      parameters:
        - $ref: '#/components/parameters/tilesetId'
      responses:
        '200':
          description: |-
            The tileset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tileSet'
        '404':
          description: |-
            tileset not found
    put:
      tags:
        - Tilesets
      summary: |-
        replace a tileset or add a new tileset
      description: |-
        Replace an existing tileset with the id `tilesetId`. If no
        such style exists, a new style with that id is added.
      operationId: updateTileset
      parameters:
        - $ref: '#/components/parameters/tilesetId'
      requestBody:
        description: |-
          A single tileset in one of the supported tileset encodings.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/tileSet'
      responses:
        '204':
          description: |-
            tileset updated or created successfully
    delete:
      tags:
        - Tilesets
      summary: |-
        delete a tileset
      description: |-
        Delete an existing tileset with the id `tilesetId`. If no
        such tileset exists, an error is returned.
      operationId: deleteTileset
      parameters:
        - $ref: '#/components/parameters/tilesetId'
      responses:
        '204':
          description: |-
            tileset deleted
        '404':
          description: |-
            tileset not found
  /tilesets/{tilesetId}/tiles/{tileMatrixSetId}/{tileMatrix}/{tileRow}/{tileCol}:
    get:
      tags:
        - Tilesets
      summary: fetch a tile from one or more geospatial data resources
      description: Retrieves a tile in the requested tileMatrixSet, on the requested
        tileMatrix in the TileMatrixSet, with the requested tile indices (tileRow,
        tileCol). The tile has multiple collections (formerly refered as layers) with
        all selected features in the bounding box of the tile.
      operationId: getTile
      parameters:
        - $ref: '#/components/parameters/tilesetId'
        - name: tileMatrixSetId
          in: path
          description: Identifier of a specific tiling scheme. It can be one of those
            specified in Annex D.1 of the OGC 17-083r2 standard or one defined in this
            service.
          required: true
          schema:
            type: string
          example: WebMercatorQuad
        - name: tileMatrix
          in: path
          description: |-
            Identifier selecting one of the scales (z) defined in the TileMatrixSet and
            representing the scaleDenominator the tile.
          required: true
          schema:
            type: integer
          example: "11"
        - name: tileRow
          in: path
          description: Row index (y) of the tile on the selected TileMatrix. It cannot exceed
            the MatrixWidth-1 for the selected TileMatrix.
          required: true
          schema:
            minimum: 0
            type: integer
          example: "827"
        - name: tileCol
          in: path
          description: Column index (x) of the tile on the selected TileMatrix. It cannot
            exceed the MatrixHeight-1 for the selected TileMatrix.
          required: true
          schema:
            minimum: 0
            type: integer
          example: 1231
      responses:
        "200":
          description: A Mapbox Vector Tile of the collection.
          content:
            application/vnd.mapbox-vector-tile:
              schema:
                type: string
                format: byte
        "404":
          description: The requested URI was not found
        "500":
          description: A server error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
  /studio/{entitiesKind}:
    get:
      tags:
        - Entities
      summary: the entities
      description: A list of all maps metadata available.
      operationId: getEntities
      parameters:
        - $ref: '#/components/parameters/entitiesKind'
      responses:
        "200":
          description: |-
            The entities of a certain kind shared by this API.
            This response can be references directly for every service that wants only essential information at the maps level. /{entitiesType}/entityId might return more information.
            This resource provides information about and how to access the entities.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/entities'
        "500":
          description: A server error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
    post:
      tags:
        - Entities
      summary: |-
        adds a new entity
      operationId: addEntity
      parameters:
        - $ref: '#/components/parameters/entitiesKind'
      description: |-
        Adds a entity to the entity repository.

        In case, a new entity is created, the following rules
        apply.

        If the entity submitted in the request body includes an
        identifier, that identifier will be used. If an entity with that identifier
        already exists, an error is returned.

        If no identifier can be determined from the submitted
        entity, the server will assign a new identifier to the
        entity.

        The URI of the new entity is returned in the header
        `Location`.
      responses:
        '201':
          description: |-
            entity created
          headers:
            Location:
              schema:
                type: string
              description: |-
                URI of the new entity
        '409':
          description: |-
            an entity with that id already exists
      requestBody:
        description: |-
          Entity to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/entity'
  /studio/{entitiesKind}/{entityId}:
    get:
      tags:
        - Entities
      summary: |-
        fetch an entity by id
      description: |-
        Fetches the entity with identifier `entityId`. The set of
        available entities can be retrieved at `/{entitiesKind}`.
      operationId: getEntity
      parameters:
        - $ref: '#/components/parameters/entityId'
        - $ref: '#/components/parameters/entitiesKind'
      responses:
        '200':
          description: |-
            The entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/entity'
        '404':
          description: |-
            entity not found
    put:
      tags:
        - Entities
      summary: |-
        replace an entity
      description: |-
        Replace an existing entity with the id `entityId`.
      operationId: updateEntity
      parameters:
        - $ref: '#/components/parameters/entityId'
        - $ref: '#/components/parameters/entitiesKind'
      requestBody:
        description: |-
          A single entity.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/entity'
      responses:
        '204':
          description: |-
            entity updated or created successfully
    delete:
      tags:
        - Entities
      summary: |-
        delete an entity
      description: |-
        Delete an existing entity with the id `entityId`. If no
        such entity exists, an error is returned.
      operationId: deleteMap
      parameters:
        - $ref: '#/components/parameters/entityId'
        - $ref: '#/components/parameters/entitiesKind'
      responses:
        '204':
          description: |-
            entity deleted
        '404':
          description: |-
            entity not found
...
components:
  parameters:
...
    styleId:
      name: styleId
      in: path
      description: |-
        Local identifier of a style. \
        A list of all available styles can be found
        under the /styles path.
      required: true
      schema:
        type: string
        format: uuid
    tilesetId:
      name: tilesetId
      in: path
      description: |-
        Local identifier of a tileset. \
        A list of all available tilesets can be found
        under the /styles path.
      required: true
      schema:
        type: string
        format: uuid
    entityId:
      name: entityId
      in: path
      description: |-
        Local identifier of an entity. \
        A list of all available entities can be found
        under the /{entitiesType} path.
      required: true
      schema:
        type: string
        format: uuid
    entitiesKind:
      name: entitiesKind
      in: path
      description: |-
        Local identifier of an entity kind. \
        A list of all available entities can be found
        under the /{entitiesKind} path.
      required: true
      schema:
        type: string
  schemas:
...
    entity:
      type: object
      required:
        - id
      properties:
        id:
          description: identifier of the entity used, for example, in URIs
          type: string
          format: uuid
      additionalProperties: { }
    entities:
      type: array
      items:
        $ref: '#/components/schemas/entity'
    exception:
      type: object
      required:
        - code
      properties:
        code:
          type: string
        description:
          type: string
...
    mb-style:
      required:
        - layers
        - sources
        - version
      type: object
      properties:
        version:
          type: number
          example: 8
        name:
          type: string
          example: night
        sources:
          type: object
          additionalProperties:
            type: object
            properties:
              type:
                type: string
                example: vector
              url:
                type: string
                example: 'https://services.interactive-instruments.de/vtp/daraa/tiles/default/{z}/{y}/{x}?f=mvt'
        sprite:
          type: string
          example: 'http://vtp2018.s3-eu-west-1.amazonaws.com/static/{entitiesType}torestyle/sprites/sprites'
        glyphs:
          type: string
          example: 'http://fonts.openmaptiles.org/{fontstack}/{range}.pbf'
        layers:
          type: array
          items:
            $ref: '#/components/schemas/layers-array'
        center:
          type: array
          items:
            type: number
        metadata:
          type: object
        zoom:
          type: number
    layers-array:
      required:
        - id
        - type
      type: object
      properties:
        id:
          type: string
          example: '1'
        type:
          type: string
          example: fill
          enum:
            - fill
            - line
            - symbol
            - circle
            - heatmap
            - fill-extrusion
            - raster
            - hillshade
            - background
        filter:
          type: array
          items: {}
        source:
          type: string
          example: daraa
        source-layer:
          type: string
          example: vegetationsrf
        layout:
          type: object
        minzoom:
          type: number
        maxzoom:
          type: number
        paint:
          type: object
...
    tileSet:
      type: object
      required:
        - tilejson
        - tiles
      properties:
        tilejson:
          type: string
          pattern: "\\d+\\.\\d+\\.\\d+\\w?[\\w\\d]*"
        name:
          type: string
        description:
          type: string
        version:
          type: string
          pattern: "\\d+\\.\\d+\\.\\d+\\w?[\\w\\d]*"
        attribution:
          type: string
        template:
          type: string
        legend:
          type: string
        scheme:
          type: string
        tiles:
          type: array
          items:
            type: string
        grids:
          type: array
          items:
            type: string
        data:
          type: array
          items:
            type: string
        minzoom:
          minimum: 0
          maximum: 30
          type: integer
        maxzoom:
          minimum: 0
          maximum: 30
          type: integer
        bounds:
          type: array
          items:
            type: number
            format: double
        center:
          type: array
          items:
            type: number
        vector_layers:
          type: array
          items:
            $ref: '#/components/schemas/layer'
    layer:
      type: object
      properties:
        id:
          type: string
        description:
          type: string
        queries:
          type: array
          items:
            $ref: '#/components/schemas/query'
    query:
      type: object
      properties:
        minzoom:
          type: integer
        maxzoom:
          type: integer
        sql:
          type: string